HIVE-27022 - Split removeDuplicateCompletedTxnComponents away from AcidHouseKeeper and onto a separate timer
